Finish DoctorBooking scheduling
