+++ src/components/App.js
@@ -18,6 +18,7 @@ import DoctorSearch from './pages/DoctorSearch';
 import AdminPanel from './pages/AdminPanel';
 import DoctorsList from './pages/DoctorsList';
 import DoctorDetail from './pages/DoctorDetail';
+import DoctorBooking from './pages/DoctorBooking';
 
 class App extends Component {
   componentDidMount() {
@@ -38,6 +39,7 @@ class App extends Component {
           <Route exact path="/admin" component={AdminPanel} />
           <Route exact path="/doctors" component={DoctorsList} />
           <Route exact path="/doctors/:id" component={DoctorDetail} />
+          <Route exact path="/doctors/:id/book" component={DoctorBooking} />
         </Switch>
       </Fragment>
     );
